Add a libmesh for solving a generalized eigenvalue problem
This eigen system is a discretization of a Fredholm Integral Equation of the second kind.
Due to the nature of the integral equation, the sparsity pattern of the system has to
be augmented. SLEPSc is used to solve the eigen system. Eigenpairs are written to exodus.
Each timestep contains one eigenpair. The eigenvector is stored as a field variable.
The eigenvalue is stored as a global variable. This example is setup such that it can be
easily modified to accommodate more complicated geometry and different covariance functions.